Linear Least-Squares Method
Use the Linear Least-Squares Method to create an equation that fits into a polynomial or non-polynomial form. In this program, a set of data is fit to the equation [ y = a*ln(x) + b*sin(x) + c ]. The program solves for [a, b, c] and the user may use the equation to calculate the temperature (y) at any time (x).