Trapezoid Rule VS Romberg Rule (Integration)
This program uses the Trapezoid Rule and the Romberg Rule to calculate the value of the integration of a function. These methods can be used to find the value of a functions integral without performing complex integrations. The returned values are then compared to show their ability to return very similar values.
The integration function: f(x) = 2/(pi)^0.5 * ( (~integral from 0 to x~) [e^(-t)^2] )
** User inputs x **