System of Differential Equations
This program solves a System of Differential Equations using the Runge-Kutta 4th Order method with a 3xN array. After the calculations are carried out over the given range, the smallest values for 't' (minutes) are given to optimize the time to fill the volume of the tank. 
(The context of the problem is from "Numerical Mathematics and Computing; 6e" - Cheney & Kincaid.